+++ diesel/src/r2d2.rs
@@ -17,7 +17,10 @@ use std::fmt;
 use std::marker::PhantomData;
 
 use crate::backend::Backend;
-use crate::connection::{ConnectionGatWorkaround, SimpleConnection, TransactionManager};
+use crate::connection::commit_error_processor::{CommitErrorOutcome, CommitErrorProcessor};
+use crate::connection::{
+    ConnectionGatWorkaround, SimpleConnection, TransactionManager, TransactionManagerStatus,
+};
 use crate::expression::QueryMetadata;
 use crate::prelude::*;
 use crate::query_builder::{AsQuery, QueryFragment, QueryId};
@@ -77,26 +80,16 @@ impl ::std::error::Error for Error {}
 pub trait R2D2Connection: Connection {
     /// Check if a connection is still valid
     fn ping(&mut self) -> QueryResult<()>;
-}
 
-#[cfg(feature = "postgres")]
-impl R2D2Connection for crate::pg::PgConnection {
-    fn ping(&mut self) -> QueryResult<()> {
-        self.execute("SELECT 1").map(|_| ())
-    }
-}
-
-#[cfg(feature = "mysql")]
-impl R2D2Connection for crate::mysql::MysqlConnection {
-    fn ping(&mut self) -> QueryResult<()> {
-        self.execute("SELECT 1").map(|_| ())
-    }
-}
-
-#[cfg(feature = "sqlite")]
-impl R2D2Connection for crate::sqlite::SqliteConnection {
-    fn ping(&mut self) -> QueryResult<()> {
-        self.execute("SELECT 1").map(|_| ())
+    /// Checks if the connection is broken and should not be reused
+    ///
+    /// This method should return only contain a fast non-blocking check
+    /// if the connection is considered to be broken or not. See
+    /// [ManageConnection::has_broken] for details.
+    ///
+    /// The default implementation does not consider any connection as broken
+    fn is_broken(&mut self) -> bool {
+        false
     }
 }
 
@@ -115,8 +108,8 @@ where
         conn.ping().map_err(Error::QueryError)
     }
 
-    fn has_broken(&self, _conn: &mut T) -> bool {
-        std::thread::panicking()
+    fn has_broken(&self, conn: &mut T) -> bool {
+        std::thread::panicking() || conn.is_broken()
     }
 }
 
@@ -138,6 +131,16 @@ where
 {
     type Cursor = <M::Connection as ConnectionGatWorkaround<'conn, 'query, DB>>::Cursor;
     type Row = <M::Connection as ConnectionGatWorkaround<'conn, 'query, DB>>::Row;
+}
+
+impl<M> CommitErrorProcessor for PooledConnection<M>
+where
+    M: ManageConnection,
+    M::Connection: R2D2Connection + CommitErrorProcessor + Send + 'static,
+{
+    fn process_commit_error(&self, error: crate::result::Error) -> CommitErrorOutcome {
+        (&**self).process_commit_error(error)
+    }
 }
 
 impl<M> Connection for PooledConnection<M>
@@ -213,8 +216,10 @@ where
         T::commit_transaction(&mut **conn)
     }
 
-    fn get_transaction_depth(conn: &mut PooledConnection<M>) -> u32 {
-        T::get_transaction_depth(&mut **conn)
+    fn transaction_manager_status_mut(
+        conn: &mut PooledConnection<M>,
+    ) -> &mut TransactionManagerStatus {
+        T::transaction_manager_status_mut(&mut **conn)
     }
 }