+++ diesel/src/result.rs
@@ -79,6 +79,25 @@ pub enum Error {
     /// Attempted to perform an operation that cannot be done inside a transaction
     /// when a transaction was already open.
     AlreadyInTransaction,
+
+    /// Attempted to perform an operation that can only be done inside a transaction
+    /// when no transaction was open
+    NotInTransaction,
+
+    /// Transaction broken, likely due to a broken connection. No other operations are possible.
+    BrokenTransaction,
+
+    /// Commiting a transaction failed
+    ///
+    /// The transaction manager will try to perform
+    /// a rollback in such cases. Indications about the success
+    /// of this can be extracted from this error variant
+    CommitTransactionFailed {
+        /// Failure message of the commit attempt
+        commit_error: Box<Error>,
+        /// Outcome of the rollback attempt
+        rollback_result: Box<QueryResult<()>>,
+    },
 }
 
 #[derive(Debug, Clone, Copy)]
@@ -297,10 +316,28 @@ impl Display for Error {
             Error::SerializationError(ref e) => e.fmt(f),
             Error::RollbackError(ref e) => e.fmt(f),
             Error::RollbackTransaction => write!(f, "The current transaction was aborted"),
+            Error::BrokenTransaction => write!(f, "The current transaction is broken"),
             Error::AlreadyInTransaction => write!(
                 f,
                 "Cannot perform this operation while a transaction is open",
             ),
+            Error::NotInTransaction => {
+                write!(f, "Cannot perform this operation outside of a transaction",)
+            }
+            Error::CommitTransactionFailed {
+                ref commit_error,
+                ref rollback_result,
+            } => {
+                write!(
+                    f,
+                    "Commiting the current transaction failed: {}",
+                    commit_error
+                )?;
+                match &**rollback_result {
+                    Ok(()) => write!(f, " Rollback attempt was succesful"),
+                    Err(e) => write!(f, " Rollback attempt failed with {}", e),
+                }
+            }
         }
     }
 }