+++ diesel/src/sqlite/connection/mod.rs
@@ -19,6 +19,9 @@ use self::raw::RawConnection;
 use self::statement_iterator::*;
 use self::stmt::{Statement, StatementUse};
 use super::SqliteAggregateFunction;
+use crate::connection::commit_error_processor::{
+    default_process_commit_error, CommitErrorOutcome, CommitErrorProcessor,
+};
 use crate::connection::*;
 use crate::deserialize::{FromSqlRow, StaticallySizedRow};
 use crate::expression::QueryMetadata;
@@ -58,6 +61,18 @@ impl SimpleConnection for SqliteConnection {
 impl<'conn, 'query> ConnectionGatWorkaround<'conn, 'query, Sqlite> for SqliteConnection {
     type Cursor = StatementIterator<'conn, 'query>;
     type Row = self::row::SqliteRow<'conn, 'query>;
+}
+
+impl CommitErrorProcessor for SqliteConnection {
+    fn process_commit_error(&self, error: Error) -> CommitErrorOutcome {
+        let state = match self.transaction_state.status {
+            TransactionManagerStatus::InError => {
+                return CommitErrorOutcome::Throw(Error::BrokenTransaction)
+            }
+            TransactionManagerStatus::Valid(ref v) => v,
+        };
+        default_process_commit_error(state, error)
+    }
 }
 
 impl Connection for SqliteConnection {
@@ -121,6 +136,21 @@ impl Connection for SqliteConnection {
         Self: Sized,
     {
         &mut self.transaction_state
+    }
+}
+
+#[cfg(feature = "r2d2")]
+impl crate::r2d2::R2D2Connection for crate::sqlite::SqliteConnection {
+    fn ping(&mut self) -> QueryResult<()> {
+        self.execute("SELECT 1").map(|_| ())
+    }
+
+    fn is_broken(&mut self) -> bool {
+        self.transaction_state
+            .status
+            .transaction_depth()
+            .map(|d| d.is_none())
+            .unwrap_or(true)
     }
 }