+++ diesel_tests/tests/errors.rs
@@ -108,7 +108,7 @@ fn foreign_key_violation_correct_constraint_name() {
 #[cfg(feature = "postgres")]
 fn isolation_errors_are_detected() {
     use diesel::result::DatabaseErrorKind::SerializationFailure;
-    use diesel::result::Error::DatabaseError;
+    use diesel::result::Error::{CommitTransactionFailed, DatabaseError};
     use std::sync::{Arc, Barrier};
     use std::thread;
 
@@ -144,10 +144,12 @@ fn isolation_errors_are_detected() {
         .execute(conn)
         .unwrap();
 
-    let barrier = Arc::new(Barrier::new(2));
+    let before_barrier = Arc::new(Barrier::new(2));
+    let after_barrier = Arc::new(Barrier::new(2));
     let threads = (1..3)
         .map(|i| {
-            let barrier = barrier.clone();
+            let before_barrier = before_barrier.clone();
+            let after_barrier = after_barrier.clone();
             thread::spawn(move || {
                 let conn = &mut connection_without_transaction();
 
@@ -157,12 +159,14 @@ fn isolation_errors_are_detected() {
                         .count()
                         .execute(conn)?;
 
-                    barrier.wait();
+                    before_barrier.wait();
 
                     let other_i = if i == 1 { 2 } else { 1 };
-                    insert_into(isolation_example::table)
+                    let r = insert_into(isolation_example::table)
                         .values(isolation_example::class.eq(other_i))
-                        .execute(conn)
+                        .execute(conn);
+                    after_barrier.wait();
+                    r
                 })
             })
         })
@@ -176,7 +180,11 @@ fn isolation_errors_are_detected() {
     results.sort_by_key(|r| r.is_err());
 
     assert_matches!(results[0], Ok(_));
-    assert_matches!(results[1], Err(DatabaseError(SerializationFailure, _)));
+    assert_matches!(results[1],
+                    Err(CommitTransactionFailed { ref commit_error, ..})
+                    if matches!(
+                        &** commit_error,
+                        DatabaseError(SerializationFailure, _)));
 }
 
 #[test]